r10720/sofa : FIX : compilation error on windows vc++ removed warnings
Former-commit-id: 032eb621e65c83141d3c95c352662bc4e7eecdad

+++ modules/sofa/component/topology/EdgeSetGeometryAlgorithms.inl
@@ -224,7 +224,7 @@ void EdgeSetGeometryAlgorithms<DataTypes>::writeMSHfile(const char *filename) co
 
     const typename DataTypes::VecCoord& vect_c = *(this->object->getX());
 
-    const unsigned int numVertices = vect_c.size();
+    const unsigned int numVertices = (unsigned int)vect_c.size();
 
     myfile << "$NOD\n";
     myfile << numVertices <<"\n";
